+++ apps/web/app/[locale]/(with-layout)/sign-up/creator/utils/signup.ts
